requirements.txt
fastapi==0.111.0
fitz==0.0.1.dev2
langchain==0.2.1
langchain_community==0.2.1
langchain_openai==0.1.7
langchain_text_splitters==0.2.0
numpy==1.25.2
openai==1.30.4
pandas==2.0.3
Pillow==10.3.0
psycopg2==2.9.7
pydantic==2.7.2
pymongo==4.6.1
pytesseract==0.3.10
python-dotenv==1.0.1
PyYAML==6.0.1
qdrant_client==1.9.1
redis==5.0.0
Requests==2.32.2
sshtunnel==0.4.0
streamlit==1.34.0
streamlit_float==0.3.4
streamlit_option_menu==0.3.12
streamlit_tags==1.2.8
streamlit-aggrid==1.0.5
tqdm==4.66.1
uvicorn==0.30.0
